Conexão feita com o SAT Control ID

+++ Views/VendaItemsSelecionados.xaml.cs
@@ -25,19 +25,9 @@ namespace FortalezaDesktop.Views
         {
             public int Iditem { get; set; }
             public int Quantidade { get; set; }
-            public ItemVendaSelectedEventArgs(int iditem, int quantidade)
+            public ItemVendaSelectedEventArgs(int iditem)
             {
                 Iditem = iditem;
-                Quantidade = quantidade;
-            }
-        }
-
-        public class CancelarItemSelectedEventArgs : EventArgs
-        {
-            public int IdItem { get; set; }
-            public CancelarItemSelectedEventArgs(int iditem)
-            {
-                IdItem = iditem;
             }
         }
 
@@ -48,12 +38,16 @@ namespace FortalezaDesktop.Views
 
         public Venda Venda { get; set; }
         public bool CriandoVenda { get; set; }
+        public bool AdicionandoProduto { get; set; }
+        public bool ReceberWasClicked { get; set; }
         public int TipoVenda { get; set; }
 
         public VendaItemsSelecionados(int tipoVenda = 0)
         {
             InitializeComponent();
+            ReceberWasClicked = false;
             CriandoVenda = false;
+            AdicionandoProduto = false;
             TipoVenda = tipoVenda;
         }
 
@@ -62,7 +56,7 @@ namespace FortalezaDesktop.Views
             Venda = await Venda.ReloadInstance(new Dictionary<string, string> {
                 {"itemvendas", "true" }
             });
-            await LoadVenda();
+            LoadVenda();
         }
 
         public async Task LoadVenda(int idvenda)
@@ -74,16 +68,17 @@ namespace FortalezaDesktop.Views
             await ReloadVenda();
         }
 
-        public async Task LoadVenda()
+        public void LoadVenda()
         {
             DataContext = null;
             DataContext = Venda;
         }
 
-        public async Task LimparVenda()
+        public void LimparVenda()
         {
             Venda = null;
-            await LoadVenda();
+            LoadVenda();
+            textboxCodigo.Focus();
         }
 
         public async Task<bool> ValidateVenda()
@@ -96,7 +91,7 @@ namespace FortalezaDesktop.Views
 
             if (Venda == null)
             {
-                Caixa caixaAberto = await (new Caixa()).GetCaixaAberto();
+                Caixa caixaAberto = await (new Caixa()).GetCaixaAberto(UserPreferences.Preferences.IdnomeCaixa);
                 if (caixaAberto != null)
                 {
                     CriandoVenda = true;
@@ -104,11 +99,13 @@ namespace FortalezaDesktop.Views
                     {
                         Tipo = TipoVenda,
                         Aberta = 1,
-                        Idresponsavel = UserControl.UsuarioLogado.Idusuario,
+                        Idresponsavel = UserController.UsuarioLogado.Idusuario,
                         Paga = 0,
-                        HoraEntrada = DateTime.UtcNow,
+                        HoraEntrada = DateTime.Now,
                         Acrescimo = 0,
                         ValorPago = 0,
+                        Idpdv = UserPreferences.Preferences.Idpdv,
+                        Idcaixa = caixaAberto.Idcaixa
                     };
                     bool success = await Venda.SaveInstance();
                     if(!success)
@@ -128,24 +125,55 @@ namespace FortalezaDesktop.Views
             return true;
         }
 
-        public async Task AddProdutoVenda(int iditem, int quantidade)
+        public async Task AddProdutoVenda(int iditem)
         {
-            if (await ValidateVenda())
+            if(!AdicionandoProduto)
             {
-                ItemVenda itemVenda = new ItemVenda
+                AdicionandoProduto = true;
+                if (await ValidateVenda())
                 {
-                    Iditem = iditem,
-                    Quantidade = quantidade
-                };
+                    int quantidade;
+                    try
+                    {
+                        quantidade = int.Parse(textboxQuantidade.Text);
+                        textboxQuantidade.Text = "1";
+                    }
+                    catch
+                    {
+                        MessageBox.Show("Quantidade inserida inválida.", "Aviso", MessageBoxButton.OK, MessageBoxImage.Information);
+                        AdicionandoProduto = false;
+                        return;
+                    }
 
-                //aviso de estoque indisponivel?
+                    ItemVenda itemVenda = new ItemVenda
+                    {
+                        Iditem = iditem,
+                        Quantidade = quantidade
+                    };
 
-                await Venda.SaveItemVenda(itemVenda);
-                await ReloadVenda();
+                    //aviso de estoque indisponivel?
+
+                    await Venda.SaveItemVenda(itemVenda);
+                    await ReloadVenda();
+                    textboxQuantidade.Text = 1.ToString();
+                    textboxCodigo.Text = "";
+                    textboxCodigo.Focus();
+                    AdicionandoProduto = false;
+                }
+                else
+                {
+                    AdicionandoProduto = false;
+                }
             }
+
         }
 
         private void ButtonIncrease_Click(object sender, RoutedEventArgs e)
+        {
+            QuantidadeIncrease();
+        }
+
+        private void QuantidadeIncrease()
         {
             try
             {
@@ -153,49 +181,112 @@ namespace FortalezaDesktop.Views
             }
             catch
             {
-
+                textboxQuantidade.Text = 2.ToString();
             }
         }
+
         private void ButtonDecrease_Click(object sender, RoutedEventArgs e)
+        {
+            QuantidadeDecrease();
+        }
+
+        private void QuantidadeDecrease()
         {
             try
             {
-                textboxQuantidade.Text = (int.Parse(textboxQuantidade.Text) - 1).ToString();
+                int current = int.Parse(textboxQuantidade.Text);
+                if (current > 1)
+                {
+                    textboxQuantidade.Text = (current - 1).ToString();
+                }
             }
             catch
             {
-
+                textboxQuantidade.Text = 1.ToString();
             }
         }
 
         private async void buttonQuantidadeOk_Click(object sender, RoutedEventArgs e)
         {
-            int quantidade;
-            int codigo;
-            
+            await SelecionarItemCodigo();
+        }
+
+        private async Task SelecionarItemCodigo()
+        {
+            if (string.IsNullOrEmpty(textboxCodigo.Text))
+            {
+                AbrirBuscaProdutos();
+                return;
+            }
+
             try
             {
-                quantidade = int.Parse(textboxQuantidade.Text);
-                codigo = int.Parse(textboxCodigo.Text);
-                ItemVendaSelected?.Invoke(this, new ItemVendaSelectedEventArgs(codigo, quantidade));
+                int codigo = int.Parse(textboxCodigo.Text);
+                bool exists = await new Item().ItemExists(codigo);
+                if (exists)
+                {
+                    ItemVendaSelected?.Invoke(this, new ItemVendaSelectedEventArgs(codigo));
+                }
+                else
+                {
+                    AbrirBuscaProdutos(textboxCodigo.Text);
+                }
             }
             catch
             {
+                AbrirBuscaProdutos(textboxCodigo.Text);
                 return;
             }
         }
 
-        private void ButtonReceber_Click(object sender, RoutedEventArgs e)
+        private async void AbrirBuscaProdutos(string busca = "")
         {
-            ReceberClicked?.Invoke(this, new EventArgs());
+            PedidoDetailsProdutos pedidoDetailsProdutos = new PedidoDetailsProdutos();
+            pedidoDetailsProdutos.ItemSelecionado += PedidoDetailsProdutos_ItemSelecionado;
+            pedidoDetailsProdutos.Closed += PedidoDetailsProdutos_Closed;
+
+            if (!string.IsNullOrEmpty(busca))
+            {
+                pedidoDetailsProdutos.textboxBuscaDescricao.Text = textboxCodigo.Text;
+            }
+
+            pedidoDetailsProdutos.ShowDialog();
         }
 
-        private void ButtonCancelar_Click(object sender, RoutedEventArgs e)
+        private async Task ConfirmarVenda()
+        {
+            if (!ReceberWasClicked)
+            {
+                ReceberWasClicked = true;
+                Caixa caixaAberto = await (new Caixa()).GetCaixaAberto(UserPreferences.Preferences.IdnomeCaixa);
+                if (caixaAberto != null)
+                {
+                    ReceberClicked?.Invoke(this, new EventArgs());
+                }
+                else
+                {
+                    MessageBox.Show("Nenhum caixa aberto.", "Aviso", MessageBoxButton.OK, MessageBoxImage.Information);
+                }
+                ReceberWasClicked = false;
+            }
+        }
+
+        private async void ButtonReceber_Click(object sender, RoutedEventArgs e)
+        {
+            await ConfirmarVenda();
+        }
+
+        private void Cancelar()
         {
             VendaCancelarItem cancelarItem = new VendaCancelarItem();
             cancelarItem.CancelarVenda += CancelarItem_CancelarVenda;
             cancelarItem.CancelarItem += CancelarItem_CancelarItem;
-            cancelarItem.Show();
+            cancelarItem.ShowDialog();
+        }
+
+        private void ButtonCancelar_Click(object sender, RoutedEventArgs e)
+        {
+            Cancelar();
         }
 
         private void CancelarItem_CancelarVenda(object sender, EventArgs e)
@@ -205,26 +296,91 @@ namespace FortalezaDesktop.Views
 
         private async void CancelarItem_CancelarItem(object sender, VendaCancelarItem.CancelarItemEventArgs e)
         {
-            if (Venda.ItemVenda != null)
+            if(Venda != null)
             {
-                List<ItemVenda> query = Venda.ItemVenda.Where(q => q.Indice == e.IndiceItemVenda).ToList();
-                if (query != null)
+                if (Venda.ItemVenda != null)
                 {
-                    if(query.Count > 0)
+                    List<ItemVenda> query = Venda.ItemVenda.Where(q => q.Indice == e.IndiceItemVenda).ToList();
+                    if (query != null)
                     {
-                        ItemVenda itemVenda = query.FirstOrDefault();
-                        await itemVenda.DeleteInstance();
-                        await Venda.RecontarItems();
-                        await ReloadVenda();
+                        if (query.Count > 0)
+                        {
+                            ItemVenda itemVenda = query.FirstOrDefault();
+                            await itemVenda.DeleteInstance();
+                            await Venda.RecontarItems();
+                            await ReloadVenda();
+                        }
                     }
                 }
             }
+        }
 
+        private void Page_Loaded(object sender, RoutedEventArgs e)
+        {
+            textboxCodigo.Focus();
         }
 
         private void SelecionarCodigo_Click(object sender, RoutedEventArgs e)
         {
+            PedidoDetailsProdutos pedidoDetailsProdutos = new PedidoDetailsProdutos();
+            pedidoDetailsProdutos.ItemSelecionado += PedidoDetailsProdutos_ItemSelecionado;
+            pedidoDetailsProdutos.Closed += PedidoDetailsProdutos_Closed;
+            pedidoDetailsProdutos.ShowDialog();
+        }
 
+        private void PedidoDetailsProdutos_Closed(object sender, EventArgs e)
+        {
+            textboxQuantidade.SelectAll();
+        }
+
+        private async void PedidoDetailsProdutos_ItemSelecionado(object sender, PedidoDetailsProdutos.ItemSelecionadoArgs e)
+        {
+            textboxCodigo.Text = e.Item.Iditem.ToString();
+            await SelecionarItemCodigo();
+        }
+
+        private void textboxQuantidade_GotFocus(object sender, RoutedEventArgs e)
+        {
+            textboxQuantidade.SelectAll();
+        }
+
+        private async void TextboxCodQuan_KeyDown(object sender, KeyEventArgs e)
+        {
+            if(e.Key == Key.Return)
+            {
+                await SelecionarItemCodigo();
+            }
+        }
+
+        private async void Page_KeyDown(object sender, KeyEventArgs e)
+        {
+            if (e.Key == Key.F12)
+            {
+                await ConfirmarVenda();
+            }
+
+            if (e.Key == Key.F11)
+            {
+                Cancelar();
+            }
+
+            if (e.Key == Key.Add)
+            {
+                QuantidadeIncrease();
+            }
+
+            if (e.Key == Key.Subtract)
+            {
+                QuantidadeDecrease();
+            }
+        }
+
+        private void textboxCodigo_PreviewTextInput(object sender, TextCompositionEventArgs e)
+        {
+            if(e.Text == "+" || e.Text == "-")
+            {
+                e.Handled = true;
+            }
         }
     }
 }
